+++ test/integration/test/snippets/overview.cpp
@@ -24,10 +24,10 @@
 #include <string>
 
 #include "test_common/ci_server.hpp"
+#include "test_integration/run_coro.hpp"
 #include "test_integration/snippets/credentials.hpp"
 #include "test_integration/snippets/describe.hpp"
 #include "test_integration/snippets/get_connection.hpp"
-#include "test_integration/snippets/run_coro.hpp"
 
 #ifdef BOOST_ASIO_HAS_CO_AWAIT
 #include <boost/asio/experimental/awaitable_operators.hpp>
